Refactor ops from Prelude to Prelude.Serial.
