+++ test/Streamly/Test/Prelude.hs
@@ -11,126 +11,23 @@ module Streamly.Test.Prelude (main) where
 
 import Control.Concurrent (myThreadId, threadDelay)
 import Control.Exception (Exception, try)
-import Control.Monad (when)
 import Control.Monad.Catch (throwM)
 import Control.Monad.Error.Class (throwError, MonadError)
 import Control.Monad.Trans.Except (runExceptT, ExceptT)
-import Data.Function ((&))
-import Data.IORef (newIORef, readIORef, writeIORef)
-import Data.Int (Int64)
 import Data.List (sort)
-import Data.Maybe (fromJust, isJust)
 #if !(MIN_VERSION_base(4,11,0))
 import Data.Semigroup (Semigroup(..))
 #endif
 import System.IO (stdout, hSetBuffering, BufferMode(LineBuffering))
 import System.Random (randomIO)
 import Test.Hspec as H
-import Test.Hspec.QuickCheck (prop)
-import Test.QuickCheck (Property, choose)
-import Test.QuickCheck.Monadic (monadicIO, pick)
 
 import Streamly.Prelude (SerialT, IsStream)
-import Streamly.Internal.Data.Time.Units
-       (AbsTime, NanoSecond64(..), toRelTime64, diffAbsTime64)
-import Streamly.Internal.Data.Time.Clock (Clock(Monotonic), getTime)
 
-import qualified Streamly.Internal.Data.Fold as FL
-import qualified Streamly.Internal.Data.Unfold as UF
-import qualified Streamly.Internal.Data.Stream.IsStream as IS
 import qualified Streamly.Prelude as S
 
 toListSerial :: SerialT IO a -> IO [a]
 toListSerial = S.toList . S.serially
-
-max_length :: Int
-max_length = 1000
-
-tenPow8 :: Int64
-tenPow8 = 10^(8 :: Int)
-
-tenPow7 :: Int64
-tenPow7 = 10^(7 :: Int)
-
-takeDropTime :: NanoSecond64
-takeDropTime = NanoSecond64 $ 5 * tenPow8
-
-checkTakeDropTime :: (Maybe AbsTime, Maybe AbsTime) -> IO Bool
-checkTakeDropTime (mt0, mt1) = do
-    let graceTime = NanoSecond64 $ 8 * tenPow7
-    case mt0 of
-        Nothing -> return True
-        Just t0 ->
-            case mt1 of
-                Nothing -> return True
-                Just t1 -> do
-                    let tMax = toRelTime64 (takeDropTime + graceTime)
-                    let tMin = toRelTime64 (takeDropTime - graceTime)
-                    let t = diffAbsTime64 t1 t0
-                    let r = t >= tMin && t <= tMax
-                    when (not r) $ putStrLn $
-                        "t = " ++ show t ++
-                        " tMin = " ++ show tMin ++
-                        " tMax = " ++ show tMax
-                    return r
-
-testTakeByTime :: IO Bool
-testTakeByTime = do
-    r <-
-          S.fold ((,) <$> FL.head <*> FL.last)
-        $ IS.takeByTime takeDropTime
-        $ S.repeatM (threadDelay 1000 >> getTime Monotonic)
-    checkTakeDropTime r
-
-testDropByTime :: IO Bool
-testDropByTime = do
-    t0 <- getTime Monotonic
-    mt1 <-
-          S.head
-        $ IS.dropByTime takeDropTime
-        $ S.repeatM (threadDelay 1000 >> getTime Monotonic)
-    checkTakeDropTime (Just t0, mt1)
-
-unfold :: Property
-unfold = monadicIO $ do
-    a <- pick $ choose (0, max_length `div` 2)
-    b <- pick $ choose (0, max_length)
-    let unf = UF.enumerateFromToIntegral b
-    ls <- S.toList $ S.unfold unf a
-    return $ ls == [a..b]
-
-unfold0 :: Property
-unfold0 = monadicIO $ do
-    a <- pick $ choose (0, max_length `div` 2)
-    b <- pick $ choose (0, max_length)
-    let unf = UF.supply (UF.enumerateFromToIntegral b) a
-    ls <- S.toList $ IS.unfold0 unf
-    return $ ls == [a..b]
-
-testFromCallback :: IO Int
-testFromCallback = do
-    ref <- newIORef Nothing
-    let stream = S.map Just (IS.fromCallback (setCallback ref))
-                    `S.parallel` runCallback ref
-    S.sum $ S.map fromJust $ S.takeWhile isJust stream
-
-    where
-
-    setCallback ref cb = do
-        writeIORef ref (Just cb)
-
-    runCallback ref = S.yieldM $ do
-        cb <-
-              S.repeatM (readIORef ref)
-                & IS.delayPost 0.1
-                & S.mapMaybe id
-                & S.head
-
-        S.fromList [1..100]
-            & IS.delayPost 0.001
-            & S.mapM_ (fromJust cb)
-        threadDelay 100000
-        return Nothing
 
 -- XXX need to test that we have promptly cleaned up everything after the error
 -- XXX We can also check the output that we are expected to get before the
@@ -275,16 +172,6 @@ parallelLoops = do
 
 main :: IO ()
 main = hspec $ H.parallel $ do
-    describe "Filtering" $ do
-        it "takeByTime" (testTakeByTime `shouldReturn` True)
-        it "dropByTime" (testDropByTime `shouldReturn` True)
-
-    describe "From Generators" $ do
-        prop "unfold" unfold
-        prop "unfold0" unfold0
-
-    it "fromCallback" $ testFromCallback `shouldReturn` (50*101)
-
     describe "Miscellaneous combined examples" mixedOps
     describe "Miscellaneous combined examples aheadly" mixedOpsAheadly
     describe "Simple MonadError and MonadThrow" simpleMonadError
