+++ test/Streamly/Test/Prelude/Serial.hs
@@ -9,7 +9,13 @@
 
 module Streamly.Test.Prelude.Serial where
 
+import Control.Concurrent ( threadDelay )
+import Control.Monad ( when )
+import Data.Function ( (&) )
+import Data.IORef ( newIORef, readIORef, writeIORef )
+import Data.Int (Int64)
 import Data.List (intercalate)
+import Data.Maybe ( isJust, fromJust )
 #if __GLASGOW_HASKELL__ < 808
 import Data.Semigroup ((<>))
 #endif
@@ -26,13 +32,18 @@ import Test.QuickCheck
     , vectorOf
     , withMaxSuccess
     )
-import Test.QuickCheck.Monadic (assert, monadicIO, run)
+import Test.QuickCheck.Monadic (assert, monadicIO, pick, run)
 import Test.Hspec as H
 
 import Streamly.Prelude
       ( SerialT, IsStream, avgRate, maxBuffer, serial, serially)
 import qualified Streamly.Prelude as S
 import qualified Streamly.Data.Fold as FL
+import qualified Streamly.Internal.Data.Unfold as UF
+import qualified Streamly.Internal.Data.Stream.IsStream as IS
+import Streamly.Internal.Data.Time.Units
+       (AbsTime, NanoSecond64(..), toRelTime64, diffAbsTime64)
+import Streamly.Internal.Data.Time.Clock (Clock(Monotonic), getTime)
 
 import Streamly.Test.Common
 import Streamly.Test.Prelude.Common
@@ -163,6 +174,95 @@ associativityCheck desc t = prop desc assocCheckProp
             assocStream <- run $ S.toList $ t $ xStream <> yStream <> zStream
             listEquals (==) infixAssocstream assocStream
 
+max_length :: Int
+max_length = 1000
+
+tenPow8 :: Int64
+tenPow8 = 10^(8 :: Int)
+
+tenPow7 :: Int64
+tenPow7 = 10^(7 :: Int)
+
+takeDropTime :: NanoSecond64
+takeDropTime = NanoSecond64 $ 5 * tenPow8
+
+checkTakeDropTime :: (Maybe AbsTime, Maybe AbsTime) -> IO Bool
+checkTakeDropTime (mt0, mt1) = do
+    let graceTime = NanoSecond64 $ 8 * tenPow7
+    case mt0 of
+        Nothing -> return True
+        Just t0 ->
+            case mt1 of
+                Nothing -> return True
+                Just t1 -> do
+                    let tMax = toRelTime64 (takeDropTime + graceTime)
+                    let tMin = toRelTime64 (takeDropTime - graceTime)
+                    let t = diffAbsTime64 t1 t0
+                    let r = t >= tMin && t <= tMax
+                    when (not r) $ putStrLn $
+                        "t = " ++ show t ++
+                        " tMin = " ++ show tMin ++
+                        " tMax = " ++ show tMax
+                    return r
+
+testTakeByTime :: IO Bool
+testTakeByTime = do
+    r <-
+          S.fold ((,) <$> FL.head <*> FL.last)
+        $ IS.takeByTime takeDropTime
+        $ S.repeatM (threadDelay 1000 >> getTime Monotonic)
+    checkTakeDropTime r
+
+testDropByTime :: IO Bool
+testDropByTime = do
+    t0 <- getTime Monotonic
+    mt1 <-
+          S.head
+        $ IS.dropByTime takeDropTime
+        $ S.repeatM (threadDelay 1000 >> getTime Monotonic)
+    checkTakeDropTime (Just t0, mt1)
+
+unfold :: Property
+unfold = monadicIO $ do
+    a <- pick $ choose (0, max_length `div` 2)
+    b <- pick $ choose (0, max_length)
+    let unf = UF.enumerateFromToIntegral b
+    ls <- S.toList $ S.unfold unf a
+    return $ ls == [a..b]
+
+unfold0 :: Property
+unfold0 = monadicIO $ do
+    a <- pick $ choose (0, max_length `div` 2)
+    b <- pick $ choose (0, max_length)
+    let unf = UF.supply (UF.enumerateFromToIntegral b) a
+    ls <- S.toList $ IS.unfold0 unf
+    return $ ls == [a..b]
+
+testFromCallback :: IO Int
+testFromCallback = do
+    ref <- newIORef Nothing
+    let stream = S.map Just (IS.fromCallback (setCallback ref))
+                    `S.parallel` runCallback ref
+    S.sum $ S.map fromJust $ S.takeWhile isJust stream
+
+    where
+
+    setCallback ref cb = do
+        writeIORef ref (Just cb)
+
+    runCallback ref = S.yieldM $ do
+        cb <-
+              S.repeatM (readIORef ref)
+                & IS.delayPost 0.1
+                & S.mapMaybe id
+                & S.head
+
+        S.fromList [1..100]
+            & IS.delayPost 0.001
+            & S.mapM_ (fromJust cb)
+        threadDelay 100000
+        return Nothing
+
 main :: IO ()
 main = hspec
     $ H.parallel
@@ -208,6 +308,10 @@ main = hspec
         serialOps   $ prop "serially iterateM" . constructWithIterateM
         serialOps $ prop "serially fromIndices" . constructWithFromIndices
         serialOps $ prop "serially fromIndicesM" . constructWithFromIndicesM
+
+        describe "From Generators" $ do
+            prop "unfold" unfold
+            prop "unfold0" unfold0
 
     describe "Functor operations" $ do
         serialOps    $ functorOps S.fromFoldable "serially" (==)
@@ -269,6 +373,10 @@ main = hspec
         serialOps    $ transformCombineOpsOrdered S.fromFoldable "serially" (==)
         serialOps    $ transformCombineOpsOrdered folded "serially" (==)
 
+        describe "Filtering" $ do
+            it "takeByTime" (testTakeByTime `shouldReturn` True)
+            it "dropByTime" (testDropByTime `shouldReturn` True)
+
     describe "Stream group and split operations" $ do
         groupSplitOps "serially"
 
@@ -289,3 +397,5 @@ main = hspec
         prop "testGroupsBySep" testGroupsBySep
     
     describe "Composed MonadThrow serially" $ composeWithMonadThrow serially
+
+    it "fromCallback" $ testFromCallback `shouldReturn` (50*101)
